refactor "ir" to "element"
