+++ compiler/element/frontend/__init__.py
@@ -1,9 +1,9 @@
 import re
 
-from compiler.ir.frontend.parser import IRParser
-from compiler.ir.frontend.printer import Printer
-from compiler.ir.frontend.transformer import IRTransformer
-from compiler.ir.node import Program
+from compiler.element.frontend.parser import IRParser
+from compiler.element.frontend.printer import Printer
+from compiler.element.frontend.transformer import IRTransformer
+from compiler.element.node import Program
 
 
 class IRCompiler:
