+++ compiler/element/props/flow.py
@@ -1,8 +1,8 @@
 from typing import Callable, Dict, List, Optional, Protocol, Sequence, Tuple, TypeVar
 
-from compiler.ir.node import *
-from compiler.ir.node import Expr, Identifier, MethodCall
-from compiler.ir.props.analyzer import (
+from compiler.element.node import *
+from compiler.element.node import Expr, Identifier, MethodCall
+from compiler.element.props.analyzer import (
     AliasAnalyzer,
     CopyAnalyzer,
     DropAnalyzer,
@@ -10,7 +10,7 @@ from compiler.ir.props.analyzer import (
     StateAnalyzer,
     WriteAnalyzer,
 )
-from compiler.ir.visitor import Visitor
+from compiler.element.visitor import Visitor
 
 
 class Property:
